+++ test/acceptance/aggregation.js
@@ -134,6 +134,100 @@ describe('aggregation', function () {
     from generate_series(0, 13) x
     `;
 
+    // Some points at corners and centers of aggregation grids (zoom=1, resolution=1)
+    //     @ = point location
+    //     --+---+---+---+---+---+---+--
+    //       |   |   |   |   |   |   |
+    //     --+---+---@---@---@---+---+--
+    //       |   |   | @ | @ |   |   |
+    //     --+---+---@---@---@---+---+--- Y = 0
+    //       |   |   | @ | @ |   |   |
+    //     --+---+---@---@---@---+---+--
+    //       |   |   |   |   |   |   |
+    //     --+---+---+---+---+---+---+--
+    //                   |
+    //                 X = 0
+    // Point identifiers (cartodb_id)
+    //     --+---+---+---+---+---+---+--
+    //       |   |   |   |   |   |   |
+    //     --+---+---7---8---9---+---+--
+    //       |   |   |12 |13 |   |   |
+    //     --+---+---4---5---6---+---+---
+    //       |   |   |10 |11 |   |   |
+    //     --+---+---1---2---3---+---+--
+    //       |   |   |   |   |   |   |
+    //     --+---+---+---+---+---+---+--
+    // Point count per aggregation cell and Z=1 tiles
+    //
+    //    Tile 0,0  -+---+---+---+- Tile 1,0
+    //       |   |   | 1 | 1 | 1 |   |
+    //     --+---+---@---@---@---+---+--
+    //       |   |   | 2 | 2 | 1 |   |
+    //     --+---+---@---@---@---+---+---
+    //       |   |   | 2 | 2 | 1 |   |
+    //     --+---+---@---@---@---+---+--
+    //       |   |   |   |   |   |   |
+    //    Tile 0, 1 -+---+---+---+- Tile 1,1
+    //
+    const POINTS_SQL_GRID = `
+        WITH params AS (
+            SELECT CDB_XYZ_Resolution(1) AS l -- cell size for Z=1 res=1
+        )
+        SELECT
+            row_number() OVER () AS cartodb_id,
+            ST_SetSRID(ST_MakePoint(x*l, y*l), 3857) AS the_geom_webmercator,
+            ST_Transform(ST_SetSRID(ST_MakePoint(x*l, y*l), 3857), 4326) AS the_geom
+            FROM params, generate_series(-1,1) x, generate_series(-1,1) y
+        UNION ALL
+        SELECT
+            row_number() OVER () + 9 AS cartodb_id,
+            ST_SetSRID(ST_MakePoint(x*l-l/2, y*l-l/2), 3857) AS the_geom_webmercator,
+            ST_Transform(ST_SetSRID(ST_MakePoint(x*l-l/2, y*l-l/2), 3857), 4326) AS the_geom
+            FROM params, generate_series(0,1) x, generate_series(0,1) y
+    `;
+
+    // Points positioned inside one cell of Z=20, X=1000000, X=1000000 (the SW corner)
+    // The center of the cell is x = 18181005.874444414, y = -18181043.94366749
+    const POINTS_SQL_CELL = `
+      SELECT
+        1 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181005.8, -18181043.9), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181005.8, -18181043.9), 3857), 4326) AS the_geom
+      UNION ALL SELECT
+        2 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181005.9, -18181044.0), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181005.9, -18181044.0), 3857), 4326) AS the_geom
+      UNION ALL SELECT
+        3 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181005.87, -18181043.94), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181005.87, -18181043.94), 3857), 4326) AS the_geom
+      UNION ALL SELECT
+        4 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181005.8, -18181043.9), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181005.8, -18181043.9), 3857), 4326) AS the_geom
+    `;
+
+    // Points positioned inside one cell of Z=20, X=1000000, X=1000000 (inner cell not on border)
+    // The center of the cell is x = 18181006.023735486, y = -18181043.794376418
+    const POINTS_SQL_CELL_INNER = `
+      SELECT
+        1 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181005.95, -18181043.8), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181005.95, -18181043.8), 3857), 4326) AS the_geom
+      UNION ALL SELECT
+        2 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181006.09, -18181043.72), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181006.09, -18181043.72), 3857), 4326) AS the_geom
+      UNION ALL SELECT
+        3 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181006.02, -18181043.79), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181006.02, -18181043.79), 3857), 4326) AS the_geom
+      UNION ALL SELECT
+        4 AS cartodb_id,
+        ST_SetSRID(ST_MakePoint(18181006.01, -18181043.75), 3857) AS the_geom_webmercator,
+        ST_Transform(ST_SetSRID(ST_MakePoint(18181006.01, -18181043.75), 3857), 4326) AS the_geom
+    `;
+
     function createVectorMapConfig (layers = [
         {
             type: 'cartodb',
@@ -1151,7 +1245,6 @@ describe('aggregation', function () {
                         type: 'cartodb',
                         options: {
                             sql: POINTS_SQL_1,
-                            resolution: 256,
                             aggregation: {
                                 threshold: 1
                             }
@@ -2162,9 +2255,9 @@ describe('aggregation', function () {
 
                                 options: {
                                     sql: POINTS_SQL_0,
-                                    resolution: 1,
                                     aggregation: {
-                                        threshold: 1
+                                        threshold: 1,
+                                        resolution: 1
                                     }
                                 }
                             }
@@ -2280,6 +2373,244 @@ describe('aggregation', function () {
 
                     });
                 });
+
+                it(`for ${placement} each aggr. cell is in a single tile`, function (done) {
+                    this.mapConfig = {
+                        version: '1.6.0',
+                        buffersize: { 'mvt': 0 },
+                        layers: [
+                            {
+                                type: 'cartodb',
+
+                                options: {
+                                    sql: POINTS_SQL_GRID,
+                                    aggregation: {
+                                        threshold: 1,
+                                        resolution: 1
+                                    }
+                                }
+                            }
+                        ]
+                    };
+                    if (placement !== 'default') {
+                        this.mapConfig.layers[0].options.aggregation.placement = placement;
+                    }
+
+                    this.testClient = new TestClient(this.mapConfig);
+
+                    this.testClient.getTile(1, 0, 0, { format: 'mvt' },  (err, res, mvt) => {
+                        if (err) {
+                            return done(err);
+                        }
+                        const tile00 = JSON.parse(mvt.toGeoJSONSync(0));
+                        this.testClient.getTile(1, 0, 1, { format: 'mvt' }, (err, res, mvt) =>  {
+                            if (err) {
+                                return done(err);
+                            }
+                            const tile01 = JSON.parse(mvt.toGeoJSONSync(0));
+                            this.testClient.getTile(1, 1, 0, { format: 'mvt' }, (err, res, mvt) =>  {
+                                if (err) {
+                                    return done(err);
+                                }
+                                const tile10 = JSON.parse(mvt.toGeoJSONSync(0));
+                                this.testClient.getTile(1, 1, 1, { format: 'mvt' }, (err, res, mvt) =>  {
+                                    if (err) {
+                                        return done(err);
+                                    }
+                                    const tile11 = JSON.parse(mvt.toGeoJSONSync(0));
+
+                                    const tile00Expected = [
+                                        { cartodb_id: 4, _cdb_feature_count: 2 },
+                                        { cartodb_id: 7, _cdb_feature_count: 1 }
+                                    ];
+                                    const tile10Expected = [
+                                        { cartodb_id: 5, _cdb_feature_count: 2 },
+                                        { cartodb_id: 6, _cdb_feature_count: 1 },
+                                        { cartodb_id: 8, _cdb_feature_count: 1 },
+                                        { cartodb_id: 9, _cdb_feature_count: 1 }
+                                    ];
+                                    const tile01Expected = [
+                                        { cartodb_id: 1, _cdb_feature_count: 2 }
+                                    ];
+                                    const tile11Expected = [
+                                        { cartodb_id: 2, _cdb_feature_count: 2 },
+                                        { cartodb_id: 3, _cdb_feature_count: 1 }
+                                    ];
+                                    const tile00Actual = tile00.features.map(f => f.properties);
+                                    const tile10Actual = tile10.features.map(f => f.properties);
+                                    const tile01Actual = tile01.features.map(f => f.properties);
+                                    const tile11Actual = tile11.features.map(f => f.properties);
+                                    const orderById = (a, b) => a.cartodb_id - b.cartodb_id;
+                                    assert.deepEqual(tile00Actual.sort(orderById), tile00Expected);
+                                    assert.deepEqual(tile10Actual.sort(orderById), tile10Expected);
+                                    assert.deepEqual(tile01Actual.sort(orderById), tile01Expected);
+                                    assert.deepEqual(tile11Actual.sort(orderById), tile11Expected);
+
+                                    done();
+                                });
+                            });
+                        });
+
+                    });
+                });
+
+                it(`for ${placement} no partially aggregated cells`, function (done) {
+                    // Use level 1 with resolution 2 tiles and buffersize 1 (half the cell size)
+                    // Only the cells completely inside the buffer are aggregated
+                    this.mapConfig = {
+                        version: '1.6.0',
+                        buffersize: { 'mvt': 1 },
+                        layers: [
+                            {
+                                type: 'cartodb',
+
+                                options: {
+                                    sql: POINTS_SQL_GRID,
+                                    aggregation: {
+                                        threshold: 1,
+                                        resolution: 2
+                                    }
+                                }
+                            }
+                        ]
+                    };
+                    if (placement !== 'default') {
+                        this.mapConfig.layers[0].options.aggregation.placement = placement;
+                    }
+
+                    this.testClient = new TestClient(this.mapConfig);
+
+                    this.testClient.getTile(1, 0, 0, { format: 'mvt' },  (err, res, mvt) => {
+                        if (err) {
+                            return done(err);
+                        }
+                        const tile00 = JSON.parse(mvt.toGeoJSONSync(0));
+                        this.testClient.getTile(1, 0, 1, { format: 'mvt' }, (err, res, mvt) =>  {
+                            if (err) {
+                                return done(err);
+                            }
+                            const tile01 = JSON.parse(mvt.toGeoJSONSync(0));
+                            this.testClient.getTile(1, 1, 0, { format: 'mvt' }, (err, res, mvt) =>  {
+                                if (err) {
+                                    return done(err);
+                                }
+                                const tile10 = JSON.parse(mvt.toGeoJSONSync(0));
+                                this.testClient.getTile(1, 1, 1, { format: 'mvt' }, (err, res, mvt) =>  {
+                                    if (err) {
+                                        return done(err);
+                                    }
+                                    const tile11 = JSON.parse(mvt.toGeoJSONSync(0));
+
+                                    const tile00Expected = [
+                                        { cartodb_id: 4, _cdb_feature_count: 3 }
+                                    ];
+                                    const tile10Expected = [
+                                        { cartodb_id: 5, _cdb_feature_count: 5 }
+                                    ];
+                                    const tile01Expected = [
+                                        { cartodb_id: 1, _cdb_feature_count: 2 }
+                                    ];
+                                    const tile11Expected = [
+                                        { cartodb_id: 2, _cdb_feature_count: 3 }
+                                    ];
+                                    const tile00Actual = tile00.features.map(f => f.properties);
+                                    const tile10Actual = tile10.features.map(f => f.properties);
+                                    const tile01Actual = tile01.features.map(f => f.properties);
+                                    const tile11Actual = tile11.features.map(f => f.properties);
+                                    const orderById = (a, b) => a.cartodb_id - b.cartodb_id;
+                                    assert.deepEqual(tile00Actual.sort(orderById), tile00Expected);
+                                    assert.deepEqual(tile10Actual.sort(orderById), tile10Expected);
+                                    assert.deepEqual(tile01Actual.sort(orderById), tile01Expected);
+                                    assert.deepEqual(tile11Actual.sort(orderById), tile11Expected);
+
+                                    done();
+                                });
+                            });
+                        });
+
+                    });
+                });
+
+                it(`for ${placement} points aggregated into corner cluster`, function (done) {
+                    // this test will fail due to !bbox! lack of accuracy if strict cell filtering is in place
+                    this.mapConfig = {
+                        version: '1.6.0',
+                        buffersize: { 'mvt': 0 },
+                        layers: [
+                            {
+                                type: 'cartodb',
+
+                                options: {
+                                    sql: POINTS_SQL_CELL,
+                                    aggregation: {
+                                        threshold: 1,
+                                        resolution: 1
+                                    }
+                                }
+                            }
+                        ]
+                    };
+                    if (placement !== 'default') {
+                        this.mapConfig.layers[0].options.aggregation.placement = placement;
+                    }
+
+                    this.testClient = new TestClient(this.mapConfig);
+
+                    this.testClient.getTile(20, 1000000, 1000000, { format: 'mvt' },  (err, res, mvt) => {
+                        if (err) {
+                            return done(err);
+                        }
+                        const tile = JSON.parse(mvt.toGeoJSONSync(0));
+                        assert.equal(tile.features.length, 1);
+                        assert.equal(tile.features[0].properties._cdb_feature_count, 4);
+                        if (placement === 'point-grid') {
+                            // check geometry x = 18181005.874444414, y = -18181043.94366749
+                            const expectedPoint = [ 163.322754576802, -83.3823797469878 ];
+                            assert.deepEqual(tile.features[0].geometry.coordinates, expectedPoint);
+                        }
+                        done();
+                    });
+                });
+
+                it(`for ${placement} points aggregated into correct cluster`, function (done) {
+                    this.mapConfig = {
+                        version: '1.6.0',
+                        buffersize: { 'mvt': 0 },
+                        layers: [
+                            {
+                                type: 'cartodb',
+
+                                options: {
+                                    sql: POINTS_SQL_CELL_INNER,
+                                    aggregation: {
+                                        threshold: 1,
+                                        resolution: 1
+                                    }
+                                }
+                            }
+                        ]
+                    };
+                    if (placement !== 'default') {
+                        this.mapConfig.layers[0].options.aggregation.placement = placement;
+                    }
+
+                    this.testClient = new TestClient(this.mapConfig);
+
+                    this.testClient.getTile(20, 1000000, 1000000, { format: 'mvt' },  (err, res, mvt) => {
+                        if (err) {
+                            return done(err);
+                        }
+                        const tile = JSON.parse(mvt.toGeoJSONSync(0));
+                        assert.equal(tile.features.length, 1);
+                        assert.equal(tile.features[0].properties._cdb_feature_count, 4);
+                        if (placement === 'point-grid') {
+                            // check geometry x = 18181006.023735486, y = -18181043.794376418
+                            const expectedPoint = [ 163.322755917907, -83.3823795924354 ];
+                            assert.deepEqual(tile.features[0].geometry.coordinates, expectedPoint);
+                        }
+                        done();
+                    });
+                });
             });
 
             ['default', 'centroid', 'point-sample', 'point-grid'].forEach(placement => {
@@ -2294,9 +2625,9 @@ describe('aggregation', function () {
 
                                 options: {
                                     sql: POINTS_SQL_PAIRS,
-                                    resolution: 1,
                                     aggregation: {
-                                        threshold: 1
+                                        threshold: 1,
+                                        resolution: 1
                                     }
                                 }
                             }
@@ -2332,9 +2663,9 @@ describe('aggregation', function () {
 
                                 options: {
                                     sql: POINTS_SQL_PAIRS,
-                                    resolution: 1,
                                     aggregation: {
-                                        threshold: 1
+                                        threshold: 1,
+                                        resolution: 1
                                     },
                                     metadata: {
                                         aggrFeatureCount: 10
@@ -2373,9 +2704,9 @@ describe('aggregation', function () {
 
                                 options: {
                                     sql: POINTS_SQL_PAIRS,
-                                    resolution: 1,
                                     aggregation: {
-                                        threshold: 1
+                                        threshold: 1,
+                                        resolution: 1
                                     },
                                     metadata: {
                                         aggrFeatureCount: 0
@@ -2414,9 +2745,9 @@ describe('aggregation', function () {
 
                                 options: {
                                     sql: POINTS_SQL_PAIRS,
-                                    resolution: 1,
                                     aggregation: {
-                                        threshold: 1
+                                        threshold: 1,
+                                        resolution: 1
                                     },
                                     metadata: {
                                         featureCount: true